--- src/main/kotlin/com/server/dpmcore/bill/bill/persentation/dto/response/CreateReceiptResponse.kt
@@ -1,3 +1,0 @@
-package com.server.dpmcore.bill.bill.persentation.dto.response
-
-class CreateReceiptResponse
