+++ src/main/kotlin/com/server/dpmcore/bill/bill/presentation/controller/BillCommandApi.kt
@@ -1,7 +1,7 @@
-package com.server.dpmcore.bill.bill.persentation.controller
+package com.server.dpmcore.bill.bill.presentation.controller
 
-import com.server.dpmcore.bill.bill.persentation.dto.request.CreateBillRequest
-import com.server.dpmcore.bill.bill.persentation.dto.response.CreateBillResponse
+import com.server.dpmcore.bill.bill.presentation.dto.request.CreateBillRequest
+import com.server.dpmcore.bill.bill.presentation.dto.response.CreateBillResponse
 import com.server.dpmcore.common.exception.CustomResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
