+++ src/main/kotlin/com/server/dpmcore/bill/bill/presentation/dto/request/CreateReceiptRequest.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.bill.bill.persentation.dto.request
+package com.server.dpmcore.bill.bill.presentation.dto.request
 
 data class CreateReceiptRequest(
     val amount: Int,
