+++ src/main/kotlin/com/server/dpmcore/bill/bill/presentation/dto/response/CreateGatheringMemberResponse.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.bill.bill.persentation.dto.response
+package com.server.dpmcore.bill.bill.presentation.dto.response
 
 data class CreateGatheringMemberResponse(
     val memberId: Long,
