+++ src/main/kotlin/com/server/dpmcore/bill/bill/presentation/dto/response/CreateReceiptResponse.kt
@@ -1,0 +1,3 @@
+package com.server.dpmcore.bill.bill.presentation.dto.response
+
+class CreateReceiptResponse
