+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceipt/presentation/mapper/ReceiptMapper.kt
@@ -1,7 +1,7 @@
 package com.server.dpmcore.gathering.gatheringReceipt.presentation.mapper
 
-import com.server.dpmcore.bill.bill.persentation.dto.request.CreateReceiptRequest
-import com.server.dpmcore.bill.bill.persentation.dto.response.CreateReceiptResponse
+import com.server.dpmcore.bill.bill.presentation.dto.request.CreateReceiptRequest
+import com.server.dpmcore.bill.bill.presentation.dto.response.CreateReceiptResponse
 import com.server.dpmcore.gathering.gatheringReceipt.domain.model.Receipt
 import com.server.dpmcore.gathering.gatheringReceiptPhoto.presentation.mapper.ReceiptPhotoMapper.toReceiptPhoto
