+++ src/main/kotlin/com/server/dpmcore/gathering/gatheringReceiptPhoto/presentation/mapper/ReceiptPhotoMapper.kt
@@ -1,6 +1,6 @@
 package com.server.dpmcore.gathering.gatheringReceiptPhoto.presentation.mapper
 
-import com.server.dpmcore.bill.bill.persentation.dto.request.CreateReceiptPhotoRequest
+import com.server.dpmcore.bill.bill.presentation.dto.request.CreateReceiptPhotoRequest
 import com.server.dpmcore.gathering.gatheringReceiptPhoto.domain.model.ReceiptPhoto
 
 object ReceiptPhotoMapper {
